update styles and intro text
