+++ app/src/app.pcss
@@ -9,6 +9,13 @@
 		url('/fonts/GeistMonoVF.woff2') format('woff2-variations');
 }
 
+@font-face {
+	font-family: 'Geist';
+	src:
+		url('/fonts/GeistVF.woff2') format('woff2 supports variations'),
+		url('/fonts/GeistVF.woff2') format('woff2-variations');
+}
+
 @layer base {
 	:root {
 		--background: 60 100% 95%;
